Move connect-inject into top-level package
